change to cargo workspace layout with shared lib, add rust webserver
