+++ .dockerignore
@@ -1,0 +1,5 @@
+**/env
+**/target
+**/node_modules
+readme-images
+.git
