+++ Cargo.toml
@@ -1,0 +1,3 @@
+[workspace]
+members = ["rust-crates/*"]
+resolver = "2"
