+++ rust-crates/event-engine/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "event_engine"
+version = "0.1.0"
+edition = "2021"
+
+
+[dependencies]
+types = { path = "../types" }
+lazy_static = "1"
+anyhow = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+csv = "1.3"
+tokio = { version = "1", features = ["full"] }
+reqwest = { version = "0.11", features = ["gzip"] }
+chrono = "0.4"
+prost = "0.12"
+redis = { version = "0.23", features = ["aio", "tokio-comp", "keep-alive"]}
+flate2 = "1"
+google-cloud-storage = "0.15.0"
+
+
+[build-dependencies]
+protobuf-src = "1"
+prost-build = "0.12"
