+++ rust-crates/event-engine/src/stream_processor/stop_detector.rs
