+++ rust-crates/types/Cargo.toml
@@ -1,8 +1,9 @@
 [package]
-name = "event_engine"
+name = "types"
 version = "0.1.0"
 edition = "2021"
 
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 lazy_static = "1"
@@ -11,14 +12,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 csv = "1.3"
 tokio = { version = "1", features = ["full"] }
-reqwest = { version = "0.11", features = ["gzip"] }
-chrono = "0.4"
-prost = "0.12"
+futures-util = "0.3"
+reqwest = { version = "0.11", features = ["gzip", "blocking"] }
+flate2 = "1"
 redis = { version = "0.23", features = ["aio", "tokio-comp", "keep-alive"]}
-flate2 = "1.0.28"
-google-cloud-storage = "0.15.0"
-
-
-[build-dependencies]
-protobuf-src = "1"
-prost-build = "0.12"
