+++ rust-crates/types/src/lib.rs
@@ -1,0 +1,13 @@
+mod metadata;
+mod vehicle;
+pub mod redis_util;
+
+pub use metadata::{
+    VehicleMetadata,
+    Stop,
+    download_metadata_table_async,
+    download_metadata_table_blocking,
+    get_trip_metadata_async,
+    get_trip_metadata_blocking
+};
+pub use vehicle::Vehicle;
