+++ rust-crates/types/src/metadata/sync_table.rs
@@ -1,0 +1,36 @@
+use std::{collections::HashMap, sync::{RwLock, Mutex}, time::{SystemTime, Duration}};
+
+use super::{VehicleMetadata, download_metadata_table_blocking};
+
+static mut TABLE: Option<RwLock<HashMap<String, VehicleMetadata>>> = None;
+static mut LAST_UPDATED: Option<Mutex<SystemTime>> = None;
+
+fn ensure_table() {
+    unsafe {
+        if let None = TABLE {
+            TABLE = Some(RwLock::new(download_metadata_table_blocking().unwrap()));
+            LAST_UPDATED = Some(Mutex::new(SystemTime::now()));
+        } else {
+            if let Some(ref last_updated) = LAST_UPDATED {
+                let mut ts = last_updated.lock().unwrap();
+                if ts.elapsed().unwrap() > Duration::from_secs(24*60*60) {
+                    *ts = SystemTime::now();
+                    if let Some(ref table) = TABLE {
+                        *table.write().unwrap() = download_metadata_table_blocking().unwrap();
+                    }
+                }
+            }
+        }
+    }
+}
+
+pub fn get_trip_metadata_blocking(trip_id: &str) -> Option<VehicleMetadata> {
+    ensure_table();
+    unsafe {
+        if let Some(ref table) = TABLE {
+            table.read().unwrap().get(trip_id).cloned()
+        } else {
+            None
+        }
+    }
+}
