+++ website/public/eventsource.js
@@ -28,3 +28,9 @@ evtSource.onmessage = (event) => {
         }
     }
 };
+
+function getMetadata(trip_id) {
+    fetch(window.location.origin + "/trip_metadata/" + trip_id)
+        .then((resp) => resp.json())
+        .then((json) => console.log(json));
+}
